+++ Craftsman/Builders/SwaggerBuilder.cs
@@ -12,11 +12,11 @@
 
     public class SwaggerBuilder
     {
-        public static void AddSwagger(string solutionDirectory, SwaggerConfig swaggerConfig, string solutionName, bool addJwtAuthentication, List<Policy> policies, string projectBaseName, IFileSystem fileSystem)
+        public static void AddSwagger(string solutionDirectory, SwaggerConfig swaggerConfig, string projectName, bool addJwtAuthentication, List<Policy> policies, string projectBaseName, IFileSystem fileSystem)
         {
             if (!swaggerConfig.IsSameOrEqualTo(new SwaggerConfig()))
             {
-                AddSwaggerServiceExtension(solutionDirectory, projectBaseName, swaggerConfig, solutionName, addJwtAuthentication, policies, fileSystem);
+                AddSwaggerServiceExtension(solutionDirectory, projectBaseName, swaggerConfig, projectName, addJwtAuthentication, policies, fileSystem);
                 WebApiAppExtensionsBuilder.CreateSwaggerWebApiAppExtension(solutionDirectory, swaggerConfig, addJwtAuthentication, projectBaseName, fileSystem);
                 UpdateWebApiCsProjSwaggerSettings(solutionDirectory, projectBaseName);
             }
@@ -60,7 +60,7 @@
             File.Move(tempPath, classPath.FullClassPath);
         }
 
-        public static void AddSwaggerServiceExtension(string solutionDirectory, string projectBaseName, SwaggerConfig swaggerConfig, string solutionName, bool addJwtAuthentication, List<Policy> policies, IFileSystem fileSystem)
+        public static void AddSwaggerServiceExtension(string solutionDirectory, string projectBaseName, SwaggerConfig swaggerConfig, string projectName, bool addJwtAuthentication, List<Policy> policies, IFileSystem fileSystem)
         {
             var classPath = ClassPathHelper.WebApiServiceExtensionsClassPath(solutionDirectory, $"SwaggerServiceExtension.cs", projectBaseName);
 
@@ -73,12 +73,12 @@
             using (var fs = fileSystem.File.Create(classPath.FullClassPath))
             {
                 var data = "";
-                data = GetSwaggerServiceExtensionText(classPath.ClassNamespace, swaggerConfig, solutionName, addJwtAuthentication, policies);
+                data = GetSwaggerServiceExtensionText(classPath.ClassNamespace, swaggerConfig, projectName, addJwtAuthentication, policies);
                 fs.Write(Encoding.UTF8.GetBytes(data));
             }
         }
 
-        public static string GetSwaggerServiceExtensionText(string classNamespace, SwaggerConfig swaggerConfig, string solutionName, bool addJwtAuthentication, List<Policy> policies)
+        public static string GetSwaggerServiceExtensionText(string classNamespace, SwaggerConfig swaggerConfig, string projectName, bool addJwtAuthentication, List<Policy> policies)
         {
             return @$"namespace {classNamespace}
 {{
@@ -96,12 +96,12 @@
 
     public static class SwaggerServiceExtension
     {{
-        {GetSwaggerServiceExtensionText(swaggerConfig, solutionName, addJwtAuthentication, policies)}
+        {GetSwaggerServiceExtensionText(swaggerConfig, projectName, addJwtAuthentication, policies)}
     }}
 }}";
         }
 
-        private static string GetSwaggerServiceExtensionText(SwaggerConfig swaggerConfig, string solutionName, bool addJwtAuthentication, List<Policy> policies)
+        private static string GetSwaggerServiceExtensionText(SwaggerConfig swaggerConfig, string projectName, bool addJwtAuthentication, List<Policy> policies)
         {
             var contactUrlLine = IsCleanUri(swaggerConfig.ApiContact.Url)
                 ? $@"
@@ -155,7 +155,7 @@
             if (swaggerConfig.AddSwaggerComments)
                 SwaggerXmlComments = $@"
 
-                config.IncludeXmlComments(string.Format(@$""{{AppDomain.CurrentDomain.BaseDirectory}}{{Path.DirectorySeparatorChar}}{solutionName}.WebApi.xml""));";
+                config.IncludeXmlComments(string.Format(@$""{{AppDomain.CurrentDomain.BaseDirectory}}{{Path.DirectorySeparatorChar}}{projectName}.WebApi.xml""));";
 
             var swaggerText = $@"public static void AddSwaggerExtension(this IServiceCollection services, IConfiguration configuration)
         {{