+++ Craftsman/Helpers/FileParsingHelper.cs
@@ -62,16 +62,16 @@
                     "Please make sure you have an `IsPrimaryKey: true` option on whichever property you want to be used as your prmary key.");
         }
 
-        public static void RunSolutionNameAssignedGuard(string solutionName)
+        public static void RunSolutionNameAssignedGuard(string projectName)
         {
-            if (solutionName == null || solutionName.Length <= 0)
+            if (projectName == null || projectName.Length <= 0)
                 throw new InvalidSolutionNameException();
         }
 
-        public static void SolutionNameDoesNotEqualEntityGuard(string solutionName, List<Entity> entities)
+        public static void SolutionNameDoesNotEqualEntityGuard(string projectName, List<Entity> entities)
         {
-            if(entities.Where(e => e.Name == solutionName).ToList().Count > 0 
-                || entities.Where(e => e.Plural == solutionName).ToList().Count > 0
+            if(entities.Where(e => e.Name == projectName).ToList().Count > 0 
+                || entities.Where(e => e.Plural == projectName).ToList().Count > 0
             )
                 throw new SolutiuonNameEntityMatchException();
         }